@@ -1,10 +1,10 @@
 name: utils
 description: This package contains utilities which are commonly used in my private projects
-version: 0.0.39
+version: 0.0.40
 homepage:
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
+  sdk: ">=3.0.5 <4.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
