@@ -45,7 +45,7 @@ class _SimpleScaffoldState extends State<SimpleScaffold> {
         ),
         title: Text(
           widget.title,
-          style: Theme.of(context).textTheme.headline6,
+          style: Theme.of(context).textTheme.titleLarge,
         ),
         actions: widget.actions,
       ),
